refactor: ImageFile의 생성자에 있는 ImageFile 이름 정하는 기능과 업로드 기능을 별도의 클래스로 분리
